requirements.txt
bs4
lxml
requests
faker
pycrypto
pycookiecheat